pyproject.toml
[tool.poetry]
name = "nhssynth"
version = "0.3.1"
description = "Synthetic data generation pipeline leveraging a Differentially Private Variational Auto Encoder assessed using a variety of metrics"
authors = ["HarrisonWilde <harrisondwilde@outlook.com>"]
maintainers = ["NHSE TDAU <england.tdau@nhs.net>"]
license = "MIT"
readme = ["README.md"]
repository = "https://github.com/nhsx/NHSSynth"
keywords = ["synthetic data", "privacy", "fairness", "machine learning"]
classifiers = [
    "Development Status :: 3 - Alpha"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/nhsx/NHSSynth/issues"
"Docs" = "https://nhsx.github.io/NHSSynth"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^2.0.1"
scikit-learn = "^1.2.2"
tqdm = "^4.65.0"
gower = "^0.1.2"
pyyaml = "^6.0"
torch = "^2.0.1"
opacus = "^1.4.0"
requests = "^2.31.0"
tornado = "^6.3.2"
sdmetrics = "^0.10.1"

[tool.poetry.scripts]
nhssynth = 'nhssynth.cli:run'

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocstrings = {extras = ["python-legacy"], version = "^0.20.0"}
mkdocs-material = "^9.1.4"
mkdocs-gen-files = "^0.4.0"
mkdocs-literate-nav = "^0.6.0"
mkdocs-section-index = "^0.3.5"
mkdocs-git-revision-date-localized-plugin = "^1.2.0"
pymdown-extensions = "^10.0.1"

[tool.poetry.group.aux]
optional = true

[tool.poetry.group.aux.dependencies]
jupyter = "^1.0.0"
notebook = "^6.5.4"
pycox = "^0.2.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
tox = "^4.0.0"

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = ["ignore::DeprecationWarning:pkg_resources"]

[tool.coverage.run]
source = ["src/nhssynth/cli", "src/nhssynth/common", "src/nhssynth/modules"]
omit = [
    "src/nhssynth/common/debugging.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
